# Disable relative step flag in adafactor