--- a/parlai/utils/fp16.py
+++ b/parlai/utils/fp16.py
@@ -554,7 +554,9 @@ class Adafactor(torch.optim.Optimizer):
         beta1=None,
         weight_decay=0.0,
         scale_parameter=True,
-        relative_step=True,
+        # relative_step=True, TODO: enable it back. This leads lr decay to 0.
+        # Since for some schdulers, they only update lr per validation step.
+        # In such cases lr will keep decay every update.
         warmup_init=False,
     ):
         defaults = dict(
@@ -565,13 +567,16 @@ class Adafactor(torch.optim.Optimizer):
             beta1=beta1,
             weight_decay=weight_decay,
             scale_parameter=scale_parameter,
-            relative_step=relative_step,
+            relative_step=False,
             warmup_init=warmup_init,
         )
         super(Adafactor, self).__init__(params, defaults)
 
     def _get_lr(self, param_group, param_state):
         rel_step_sz = param_group['lr']
+        # TODO: enable it back. This leads lr decay to 0.
+        # Since for some schdulers, they only update lr per validation step.
+        # In such cases lr will keep decay every update.
         if param_group['relative_step']:
             min_step = (
                 1e-6 * param_state['step'] if param_group['warmup_init'] else 1e-2