--- a/parlai/utils/fp16.py
+++ b/parlai/utils/fp16.py
@@ -553,10 +553,10 @@ class Adafactor(torch.optim.Optimizer):
         decay_rate=-0.8,
         beta1=None,
         weight_decay=0.0,
-        scale_parameter=True,
-        # relative_step=True, TODO: enable it back. This leads lr decay to 0.
+        # scale_parameter=True, TODO: enable it back. This leads lr decay to 0.
         # Since for some schdulers, they only update lr per validation step.
         # In such cases lr will keep decay every update.
+        # relative_step=True,
         warmup_init=False,
     ):
         defaults = dict(
@@ -566,7 +566,7 @@ class Adafactor(torch.optim.Optimizer):
             decay_rate=decay_rate,
             beta1=beta1,
             weight_decay=weight_decay,
-            scale_parameter=scale_parameter,
+            scale_parameter=False,
             relative_step=False,
             warmup_init=warmup_init,
         )